# perf: improve is_ipv4_number_valid