--- a/include/ada/checkers.h
+++ b/include/ada/checkers.h
@@ -10,7 +10,7 @@ namespace ada::checkers {
   bool ends_in_a_number(std::string_view input) noexcept;
   bool is_windows_drive_letter(std::string_view input) noexcept;
   bool is_normalized_windows_drive_letter(std::string_view input) noexcept;
-  ada_really_inline constexpr bool is_ipv4_number_valid(const std::string_view input) noexcept;
+  ada_really_inline constexpr bool is_ipv4_number_valid(std::string_view::iterator iterator_start, std::string_view::iterator iterator_end) noexcept;
 
 } // namespace ada::checkers