--- a/include/ada/helpers.h
+++ b/include/ada/helpers.h
@@ -11,7 +11,6 @@ namespace ada::helpers {
 
   std::vector<std::string> split_string_view(std::string_view input, char delimiter, bool skip_empty = true);
   ada_really_inline uint64_t string_to_uint64(std::string_view view);
-  std::string from_decimal(std::string& res, uint16_t base, uint16_t input);
 
 } // namespace ada::helpers