--- a/src/helpers.cpp
+++ b/src/helpers.cpp
@@ -19,7 +19,7 @@ namespace ada::helpers {
     return out;
   }
 
-  std::string get_state(ada::state state) {
+  ada_unused std::string get_state(ada::state state) {
     switch (state) {
       case AUTHORITY: return "Authority";
       case SCHEME_START: return "Scheme Start";
@@ -49,22 +49,6 @@ namespace ada::helpers {
     uint64_t val;
     std::memcpy(&val, view.data(), sizeof(uint64_t));
     return val;
-  }
-
-  std::string from_decimal(std::string& res, uint16_t base, uint16_t input) {
-      while (input > 0) {
-        int num = input % base;
-        if (num >= 0 && num <= 9) {
-          res += static_cast<char>(num + '0');
-        } else {
-          res += static_cast<char>(num - 10 + 'a');
-        }
-        input /= base;
-      }
-
-      std::reverse(res.begin(), res.end());
-
-      return res;
   }
 
 } // namespace ada::helpers