Before
```
BasicBench_AdaURL            7374 ns         7341 ns        94908 time/byte=9.64605ns
```

After
```
BasicBench_AdaURL            7199 ns         7189 ns        97386 time/byte=9.44628ns
```