@lemire 2023-01-07
Is this a real optimization? I do not see anything in the diff that might explain a performance boost. Note that the benchmark so we can't do a direct comparison.

(The PR itself is fine, I just do not expect it to improve the performance.)