@anonrig 2023-01-07
@lemire Reducing the number of branches in the checkers.cpp resulted in 100+ns improvement. There were some unnessary checks.