@lemire 2023-01-07
> Reducing the number of branches in the checkers.cpp resulted in 100+ns improvement. There were some unnessary checks.

Ah. Yes. I stand corrected. I see the optimization now.